requirements.txt
azure-cosmos==3.1.0
azure-functions==1.0.0b4
